backend/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
pydantic==2.11.3
pydantic-settings==2.10.1
pytest==8.3.4
httpx==0.28.1
langchain==0.3.14
openai==1.59.8
anthropic==0.42.0
pinecone-client==5.0.1
PyMuPDF==1.25.3
tiktoken==0.9.0
rank-bm25==0.2.2
twilio==9.4.1
python-multipart==0.0.20
